requirements.txt
numpy==1.21.2
pandas==1.3.2
scikit-learn==1.0.2
matplotlib==3.4.3
seaborn==0.11.2
obp==0.5.2
hydra-core==1.0.6
